+++ packages/dara-components/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## NEXT
+
+-  Internal: re-export `@codemirror` packages to solve duplication of CM dependencies across UMD bundles
+
 ## 1.16.6
 
 -  Added a new improved `UiCodeEditor` implementation into this package directly, superseding the old `@darajs/ui-code-editor` package.